cmds: Fix type cast in fb_devfs_access

+++ src/cmds/testing/fb_devfs_access/fb_devfs_access.c
@@ -53,7 +53,7 @@ int main() {
 	/* Map the device to memory */
 	fbp = (char *) mmap(0, screensize, PROT_READ | PROT_WRITE, MAP_SHARED, fbfd,
 			0);
-	if ((int) fbp == -1) {
+	if ((intptr_t) fbp == -1) {
 		perror("Error: failed to map framebuffer device to memory");
 		exit(4);
 	}
